ItemCache hits and misses must be long instead of int, otherwise may run over
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@8140 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ console/src/com/exedio/cope/console/ItemCacheCop.java
@@ -40,8 +40,8 @@ final class ItemCacheCop extends ConsoleCop
 		
 		int allLimit = 0;
 		int allLevel = 0;
-		int allHits = 0;
-		int allMisses = 0;
+		long allHits = 0;
+		long allMisses = 0;
 		int allNumberOfCleanups = 0;
 		int allItemsCleanedUp = 0;
 		Date allLastCleanup = null;
